Data plugged in. Now Interface should give feedback when req is not met. Future issue

+++ Assets/Scripts/DataUnits/ItemScriptableObjects/IItemObject.cs
@@ -13,7 +13,7 @@ namespace DataUnits.ItemScriptableObjects
         public string ItemName { get; }
         public string ItemDescription { get; }
         public int UnlockPoints { get; }
-        public float Cost { get; }
+        public int Cost { get; }
         public int ItemActions { get; }
         public Sprite ItemIcon{ get; }
         public IItemTypeStats ItemStats {get;}
